pyproject.toml
[tool.poetry]
name = "pythonproject6"
version = "0.1.0"
description = ""
authors = ["SeaCoffee <sapienti8sat@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.2"
mysqlclient = "^2.2.6"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^11.1.0"
uuid = "^1.30"
celery = "^5.4.0"
redis = "^5.2.1"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
